+++ bot/__init__.py
@@ -21,7 +21,7 @@ DB_URI = os.environ.get("DB_URI")
 
 USER_SESSION = os.environ.get("USER_SESSION")
 
-VERIFY = {}
+CHAT_DETAILS = {}
 
 logging.basicConfig(
     level=logging.INFO,
