+++ bot/plugins/settings.py
@@ -6,24 +6,25 @@ import re
 from pyrogram import Client, filters
 from pyrogram.types import InlineKeyboardButton, InlineKeyboardMarkup
 
-from bot import VERIFY # pylint: disable=import-error
+from bot import CHAT_DETAILS
+from bot.plugins.utils import admin_list
+
+
 
 @Client.on_message(filters.command(["settings"]) & filters.group, group=1)
 async def settings(bot, update):
     
     chat_id = update.chat.id
     user_id = update.from_user.id if update.from_user else None
-    global VERIFY
+    global CHAT_DETAILS
+    
+    chat_dict = CHAT_DETAILS.get(str(chat_id))
+    chat_admins = chat_dict.get("admins") if chat_dict != None else None
 
-    if VERIFY.get(str(chat_id)) == None: # Make Admin's ID List
-        admin_list = []
-        async for x in bot.iter_chat_members(chat_id=chat_id, filter="administrators"):
-            admin_id = x.user.id 
-            admin_list.append(admin_id)
-        admin_list.append(None)
-        VERIFY[str(chat_id)] = admin_list
+    if ( chat_dict or chat_admins ) == None: # Make Admin's ID List
+        chat_admins = await admin_list(chat_id, bot, update)
 
-    if not user_id in VERIFY.get(str(chat_id)): # Checks if user is admin of the chat
+    if user_id not in chat_admins:
         return
     
     bot_info = await bot.get_me()
